Move components out of app
