+++ components/module-components/tags/TagManagementControls.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
-import Button from '../../../components/reusable-components/button/Button';
-import Input from '../../../components/reusable-components/input/Input';
+import Button from '../../reusable-components/button/Button';
+import Input from '../../reusable-components/input/Input';
 import styles from './tagManagementControls.module.scss';
 
 interface TagManagementControlsProps {
