+++ components/module-components/tags/Tags.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/utils/breakpoints';
+@import '../../../app/styles/utils/breakpoints';
 
 .container {
   width: 100%;
